Linux patches to add more fallbacks to machine-id location
By default dbus look for machine-id on the following paths:

- $PREFIX/var/lib/dbus/machine-id
- /etc/machine-id (FALLBACK)

This works fine as long as the dbus package is used in a conda
environment.
But a problem arises when someone tries to ship this dbus package with
an application (e.g. a Qt5 application):

- $PREFIX/var/lib/dbus/machine-id does not exist as it would be inside a
conda env (customer has no conda)
- /etc/machine-id exists only on systems with systemd

Systems without systemd will normally have this file in
/var/lib/dbus/machine-id, so it seems reasonable to use this file as
another fallback.

Last, on systems without dbus, it will fallback to
/proc/sys/kernel/random/boot_id